MessageStorageServicvice now casts Connection ID to Long

+++ plugins/frontlinesms-core/grails-app/services/frontlinesms2/MessageStorageService.groovy
@@ -8,7 +8,7 @@ class MessageStorageService implements Processor {
 		def message = x.in.body
 		assert message instanceof TextMessage
 		message = message.id ? TextMessage.findById(message.id) : message
-		message.connectionId = x.in.headers[Fconnection.HEADER_FCONNECTION_ID]
+		message.connectionId = x.in.headers[Fconnection.HEADER_FCONNECTION_ID] as Long
 		message.save(flush:true)
 	}
 }
